Make Institution id a String, for future compatibility
In 2018, an institution's id will be the LEI, which is alphanumeric, not
an integer.

+++ persistence/src/main/scala/hmda/persistence/demo/DemoInstitutions.scala
@@ -10,7 +10,7 @@ object DemoInstitutions extends ResourceUtils {
 
     lines.map { line =>
       val values = line.split('|').map(_.trim)
-      val id = values(0).toInt
+      val id = values(0)
       val name = values(1)
       val externalId = values(2)
       val externalIdType = toExternalIdType(values(3))
